+++ package.json
@@ -23,6 +23,7 @@
 	],
 	"dependencies": {
 		"@slippi/slippi-js": "^6.5.0",
+		"axios": "^1.2.1",
 		"chokidar": "^3.5.3",
 		"electron-context-menu": "^3.3.0",
 		"electron-reloader": "^1.2.3",
